Use AppVeyor to build windows wheel (#188)
* Add AppVeyor support to build windows wheel
* Fix test_limits on 32bit environments
* Ignore Python35-x64 test fail for now
  Should be fixed in next version.

+++ msgpack/_packer.pyx
@@ -271,7 +271,7 @@ cdef class Packer(object):
         msgpack_pack_ext(&self.pk, typecode, len(data))
         msgpack_pack_raw_body(&self.pk, data, len(data))
 
-    def pack_array_header(self, size_t size):
+    def pack_array_header(self, long long size):
         if size > (2**32-1):
             raise PackValueError
         cdef int ret = msgpack_pack_array(&self.pk, size)
@@ -284,7 +284,7 @@ cdef class Packer(object):
             self.pk.length = 0
             return buf
 
-    def pack_map_header(self, size_t size):
+    def pack_map_header(self, long long size):
         if size > (2**32-1):
             raise PackValueError
         cdef int ret = msgpack_pack_map(&self.pk, size)
